requirements.txt
fastapi[all]
pydantic
pydantic-settings
debugpy
asyncpg
passlib[bcrypt]
sqlmodel
# psycopg2[asyncpg]
sqlalchemy
PyJWT
greenlet
pytest
anyio
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted
fastapi-mail[all]
tortoise-orm[asyncpg]
aerich
tomlkit

